+++ packages/esm-patient-medications-app/src/medications-summary/medications-summary.component.tsx
@@ -3,8 +3,8 @@ import FloatingOrderBasketButton from './floating-order-basket-button.component'
 import MedicationsDetailsTable from '../components/medications-details-table.component';
 import { DataTableSkeleton } from 'carbon-components-react';
 import { useTranslation } from 'react-i18next';
-import { usePatientOrders } from '../utils/use-current-patient-orders.hook';
 import { EmptyState, ErrorState } from '@openmrs/esm-patient-common-lib';
+import { usePatientOrders } from '../api/api';
 
 export interface MedicationsSummaryProps {
   patientUuid: string;
