+++ packages/esm-patient-medications-app/src/medications/active-medications.component.tsx
@@ -2,12 +2,12 @@ import React from 'react';
 import MedicationsDetailsTable from '../components/medications-details-table.component';
 import styles from './active-medications.scss';
 import { useTranslation } from 'react-i18next';
-import { usePatientOrders } from '../utils/use-current-patient-orders.hook';
 import { Provider } from 'unistore/react';
 import { orderBasketStore } from './order-basket-store';
 import { DataTableSkeleton } from 'carbon-components-react';
 import { EmptyState, ErrorState } from '@openmrs/esm-patient-common-lib';
 import { attach } from '@openmrs/esm-framework';
+import { usePatientOrders } from '../api/api';
 
 interface ActiveMedicationsProps {
   patientUuid: string;
@@ -18,6 +18,7 @@ const ActiveMedications: React.FC<ActiveMedicationsProps> = ({ patientUuid, show
   const { t } = useTranslation();
   const displayText = t('activeMedications', 'Active medications');
   const headerTitle = t('activeMedications', 'active medications');
+
   const { data: activePatientOrders, isError, isLoading, isValidating } = usePatientOrders(patientUuid, 'ACTIVE');
 
   const launchOrderBasket = React.useCallback(() => {
