+++ packages/esm-patient-medications-app/translations/en.json
@@ -39,7 +39,6 @@
   "pastMedications": "Past Medications",
   "patientInstructions": "Patient Instructions",
   "patientInstructionsPlaceholder": "Additional dosing instructions (e.g. \"Take after eating\")",
-  "pillsDispensed": "Pills dispensed",
   "pillsToDispense": "Pills to dispense",
   "prescriptionDuration": "2. Prescription Duration",
   "prescriptionRefills": "Prescription Refills",
@@ -47,7 +46,6 @@
   "prnReason": "P.R.N. Reason",
   "prnReasonPlaceholder": "Reason to take medicine",
   "quantity": "Quantity",
-  "quantityDispensed": "Quantity Dispensed",
   "refills": "Refills",
   "removeFromBasket": "Remove from basket",
   "renewedOrders": "{count} order(s) being renewed (continued)",
